+++ module/vo/xray.go
@@ -2,5 +2,5 @@ package vo
 
 type XrayStatsVo struct {
 	Name  string `json:"name"`
-	Value int64  `json:"value"`
+	Value int    `json:"value"`
 }
